# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the Green Ecolution Management System.",
        "title": "Green Space Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Green Ecolution",
            "url": "https://green-ecolution.de",
            "email": "info@green-ecolution.de"
        },
        "license": {
            "name": "AGPL",
            "url": "https://raw.githubusercontent.com/green-ecolution/green-ecolution/develop/LICENSE"
        },
        "version": "develop"
    },
    "paths": {
...
        "/v1/vehicle": {
            "get": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Get all vehicles",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle"
                ],
                "summary": "Get all vehicles",
                "operationId": "get-all-vehicles",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Vehicle Type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Provider",
                        "name": "provider",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "With archived vehicles",
                        "name": "archived",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/VehicleList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Create vehicle",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle"
                ],
                "summary": "Create vehicle",
                "operationId": "create-vehicle",
                "parameters": [
                    {
                        "description": "Vehicle Create Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/VehicleCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Vehicle"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/vehicle/archive": {
            "get": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Get archived vehicle",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle"
                ],
                "summary": "Get archived vehicle",
                "operationId": "get-archive-vehicle",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Vehicle"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/vehicle/archive/{id}": {
            "post": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Archive vehicle",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle"
                ],
                "summary": "Archive vehicle",
                "operationId": "archive-vehicle",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vehicle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/vehicle/plate/{plate}": {
            "get": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Get vehicle by plate",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle"
                ],
                "summary": "Get vehicle by plate",
                "operationId": "get-vehicle-by-plate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle plate number",
                        "name": "plate",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Vehicle"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/vehicle/{id}": {
            "get": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Get vehicle by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle"
                ],
                "summary": "Get vehicle by ID",
                "operationId": "get-vehicle-by-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vehicle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Vehicle"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Update vehicle",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle"
                ],
                "summary": "Update vehicle",
                "operationId": "update-vehicle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Vehicle Update Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/VehicleUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Vehicle"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Delete vehicle",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle"
                ],
                "summary": "Delete vehicle",
                "operationId": "delete-vehicle",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vehicle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "HTTPError": {
            "type": "object",
            "required": [
                "code",
                "error",
                "method",
                "path"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "Pagination": {
            "type": "object",
            "required": [
                "current_page",
                "next_page",
                "prev_page",
                "total_pages",
                "total_records"
            ],
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "next_page": {
                    "type": "integer"
                },
                "prev_page": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_records": {
                    "type": "integer"
                }
            }
        },
...
        "Vehicle": {
            "type": "object",
            "required": [
                "created_at",
                "description",
                "driving_license",
                "height",
                "id",
                "length",
                "model",
                "number_plate",
                "status",
                "type",
                "updated_at",
                "water_capacity",
                "weight",
                "width"
            ],
            "properties": {
                "additional_information": {
                    "type": "object",
                    "additionalProperties": true
                },
                "archived_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "driving_license": {
                    "$ref": "#/definitions/DrivingLicense"
                },
                "height": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "length": {
                    "type": "number"
                },
                "model": {
                    "type": "string"
                },
                "number_plate": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/VehicleStatus"
                },
                "type": {
                    "$ref": "#/definitions/VehicleType"
                },
                "updated_at": {
                    "type": "string"
                },
                "water_capacity": {
                    "type": "number"
                },
                "weight": {
                    "type": "number"
                },
                "width": {
                    "type": "number"
                }
            }
        },
        "VehicleCreate": {
            "type": "object",
            "required": [
                "description",
                "driving_license",
                "height",
                "length",
                "model",
                "number_plate",
                "status",
                "type",
                "water_capacity",
                "weight",
                "width"
            ],
            "properties": {
                "additional_information": {
                    "type": "object",
                    "additionalProperties": true
                },
                "description": {
                    "type": "string"
                },
                "driving_license": {
                    "$ref": "#/definitions/DrivingLicense"
                },
                "height": {
                    "type": "number"
                },
                "length": {
                    "type": "number"
                },
                "model": {
                    "type": "string"
                },
                "number_plate": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/VehicleStatus"
                },
                "type": {
                    "$ref": "#/definitions/VehicleType"
                },
                "water_capacity": {
                    "type": "number"
                },
                "weight": {
                    "type": "number"
                },
                "width": {
                    "type": "number"
                }
            }
        },
...
        "VehicleList": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Vehicle"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/Pagination"
                }
            }
        },
...
        "VehicleUpdate": {
            "type": "object",
            "required": [
                "description",
                "driving_license",
                "height",
                "length",
                "model",
                "number_plate",
                "status",
                "type",
                "water_capacity",
                "weight",
                "width"
            ],
            "properties": {
                "additional_information": {
                    "type": "object",
                    "additionalProperties": true
                },
                "description": {
                    "type": "string"
                },
                "driving_license": {
                    "$ref": "#/definitions/DrivingLicense"
                },
                "height": {
                    "type": "number"
                },
                "length": {
                    "type": "number"
                },
                "model": {
                    "type": "string"
                },
                "number_plate": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/VehicleStatus"
                },
                "type": {
                    "$ref": "#/definitions/VehicleType"
                },
                "water_capacity": {
                    "type": "number"
                },
                "weight": {
                    "type": "number"
                },
                "width": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Keycloak": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://auth.green-ecolution.de/realms/green-ecolution-dev/protocol/openid-connect/auth",
            "tokenUrl": "https://auth.green-ecolution.de/realms/green-ecolution-dev/protocol/openid-connect/token"
        }
    }
}